CODE RUB: Move LabCommand Namespace
