+++ DMX.Gatekeeper.Api/Brokers/DmxApis/IDmxApiBroker.LabCommands.cs
@@ -3,12 +3,12 @@
 // ---------------------------------------------------------------
 
 using System.Threading.Tasks;
-using DMX.Gatekeeper.Api.Models.Commands;
+using DMX.Gatekeeper.Api.Models.LabCommands;
 
 namespace DMX.Gatekeeper.Api.Brokers.DmxApis
 {
     public partial interface IDmxApiBroker
     {
-        ValueTask PostCommand(LabCommand labCommand);
+        ValueTask<LabCommand> PostLabCommandAsync(LabCommand labCommand);
     }
 }
