+++ DMX.Gatekeeper.Api/Models/LabCommands/CommandStatus.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ---------------------------------------------------------------
 
-namespace DMX.Gatekeeper.Api.Models.Commands
+namespace DMX.Gatekeeper.Api.Models.LabCommands
 {
     public enum CommandStatus
     {
