+++ DMX.Gatekeeper.Api/Models/LabCommands/CommandType.cs
@@ -2,11 +2,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ---------------------------------------------------------------
 
-namespace DMX.Gatekeeper.Api.Models.Commands
+namespace DMX.Gatekeeper.Api.Models.LabCommands
 {
     public enum CommandType
     {
+        Execute,
         Download,
-        Execute
     }
 }
